Lazy load details tabs content
