+++ www/front_src/src/Resources/Details/LoadingSkeleton.tsx
@@ -1,0 +1,31 @@
+import * as React from 'react';
+
+import { makeStyles, styled } from '@material-ui/core';
+
+import { LoadingSkeleton } from '@centreon/ui';
+
+const useStyles = makeStyles((theme) => ({
+  loadingSkeleton: {
+    display: 'grid',
+    gridTemplateRows: 'repeat(3, 67px)',
+    rowGap: theme.spacing(2),
+  },
+}));
+
+const CardSkeleton = styled(LoadingSkeleton)(() => ({
+  transform: 'none',
+}));
+
+const DetailsLoadingSkeleton = (): JSX.Element => {
+  const classes = useStyles();
+
+  return (
+    <div className={classes.loadingSkeleton}>
+      <CardSkeleton height="100%" />
+      <CardSkeleton height="100%" />
+      <CardSkeleton height="100%" />
+    </div>
+  );
+};
+
+export default DetailsLoadingSkeleton;
